Use indexed-traversable
