+++ lens.cabal
@@ -193,12 +193,12 @@ library
     bifunctors                >= 5.1      && < 6,
     bytestring                >= 0.10.4.0 && < 0.12,
     call-stack                >= 0.1      && < 0.3,
-    comonad                   >= 4        && < 6,
+    comonad                   >= 5.0.7    && < 6,
     contravariant             >= 1.3      && < 2,
     containers                >= 0.5.5.1  && < 0.7,
     distributive              >= 0.3      && < 1,
     filepath                  >= 1.2.0.0  && < 1.5,
-    free                      >= 4        && < 6,
+    free                      >= 5.1.5    && < 6,
     ghc-prim,
     hashable                  >= 1.1.2.3  && < 1.4,
     kan-extensions            >= 5        && < 6,
@@ -217,7 +217,9 @@ library
     transformers              >= 0.3.0.0  && < 0.6,
     transformers-compat       >= 0.4      && < 1,
     unordered-containers      >= 0.2.4    && < 0.3,
-    vector                    >= 0.9      && < 0.13
+    vector                    >= 0.9      && < 0.13,
+    indexed-traversable           >= 0.1      && < 0.2,
+    indexed-traversable-instances >= 0.1      && < 0.2
 
   if !impl(ghc >= 8.0)
     build-depends:
